+++ src/Rocks.Analysis.IntegrationTests/MockDefinitions.cs
@@ -36,6 +36,8 @@
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.ArgTestTypes.IHaveArgument), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AsynchronousTestTypes.AsyncEnumeration), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AsynchronousTestTypes.IAmAsynchronous), Rocks.BuildType.Create | Rocks.BuildType.Make)]
+[assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AbstractClassConstructorTestTypes.AbstractClassConstructor), Rocks.BuildType.Create | Rocks.BuildType.Make)]
+[assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AttributeTestTypes.UsingMemberNotNullWhen), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AttributeTestTypes.ConventionDispatcher), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AttributeTestTypes.ITense), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.AttributeTestTypes.NotNullIfNotCases), Rocks.BuildType.Create | Rocks.BuildType.Make)]